+++ db/updates/update_3.9-4.0.sql
@@ -1,1 +1,2 @@
 ALTER TABLE `formulasMetaData` ADD `customer_id` INT NOT NULL DEFAULT '0' AFTER `madeOn`;
+ALTER TABLE `ingredients` ADD `einecs` VARCHAR(255) NULL AFTER `cas`;