+++ releasenotes.md
@@ -3,3 +3,4 @@ Whats New:
 Major improvements in ingredients page
 You can now assign a formula to a customer
 No need to create an account to retrieve ingredients from PV Online
+Added EINECS support in ingredients